refactor: remove vanilla directory
