+++ packages/lit/index.html
@@ -7,7 +7,7 @@
     <title>@kotice/lit-web-components-sandbox</title>
     <link rel="stylesheet" href="./src/assets/globals.css" />
     <script type="module" src="/src/my-app.ts"></script>
-    <script type="module" src="/src/components/lit/my-search-input.ts"></script>
+    <script type="module" src="/src/components/my-search-input.ts"></script>
   </head>
   <body>
     <my-app>
