+++ packages/lit/src/components/my-search-input.ts
